+++ lib/event_source.dart
@@ -1,0 +1,4 @@
+/// Helps tweak [EventSource] requests outside browsers.
+library universal_html.event_source;
+
+import 'html.dart';